core: move clusterInfo code to the controller package
The CSI package needs to load clusterInfo, today this code is in the mon
package which makes the call of LoadClusterInfo impossible without
having a circular import.

(cherry picked from commit 583791c45c7a752dacb81f0f27533cae8a586fa5)
